+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #67 JJAVA_STARTUP_SCRIPT can't start up spark
 * #102 Eval startup snippets explicitly
 * #105 Exploding classpath
+* #107 "java.class.path" uses an incorrect separator
 
 ## 1.0-a6